Use media queries for the radio track layout :(
I was not able to make what I wanted with the Every Layout components.
Using a sidebar element within an intrinsically centered element has a
problem: when the title of the track is very short, its width is less
than 50% of the [cover|track-info] block, so it wraps, even though it
would have room to be displayed side-by-side with the cover.

I might have missed something though.

Also, media queries make it easy to adjust the spacing flow on small
screen, where we want smaller margins so that everything fits inside the
viewport while still having lots of white space on bigger screens.